agents/vendor_risk_analysis_agent/requirements.txt
google-adk
pydantic
vertexai
requests
uvicorn
fastapi
python-dotenv
beautifulsoup4
